requirements.txt
pyyaml==6.0.3
reportlab==4.4.9
openai==2.15.0
beautifulsoup4==4.14.3
requests==2.32.5
pdfplumber==0.11.9
